argument typo `-p`
```
-p portnum,locport (*)
Direct the control connection to a netserver listening on the specified port, rather than using a "netperf" entry in /etc/services or the internal default (port 12865). If ",locport" is specified the control connection will be established from that local port number. Specifying a single port number with no comma will specify only the remote netserver port number and will leave local port number selection to the stack.
-P 0|1
Show (1) or suppress (0) the test banner.
```